+++ CHANGELOG.md
@@ -89,6 +89,7 @@ a release.
 
 ### Added
 - Tree: Add `Nested::ALLOWED_NODE_POSITIONS` constant in order to expose the available node positions
+- Tree: Add pre-order and level-order traversals (#2498). Add `getNextNode()` and `getNextNodes()` methods to traverse the tree
 - Support for `doctrine/collections` 2.0
 - Support for `doctrine/event-manager` 2.0
 - Loggable: Add `LogEntryInterface` interface in order to be implemented by log entry models